More threading shinanagans.
Really cleaned up what I did in my previous commit. Created a new WorkQueueThread
class that encapsulates all of the work queue stuff for a single dedicated thread.

I replaced the Tasks I was using in the FileSystem class to refresh the file system
with a singular new WorkQueueThead object.