Some refactor, put things in more sensible namespaces.
Added IAutoThemeUser and started using the AutoThemeNodes group to subscribe
nodes to top level AutoTheme changes and provide access to the app theme.
Do not try to access the AutoTheme in _Ready() this way, since the Theme won't
be populated. Let the interface implementation handle it.

My Main node in Program.tscn automatically hooks up all nodes in AutoThemeNodes group.